+++ RELEASE-NOTES.md
@@ -14,6 +14,7 @@
 ## Fixes and improvements
 * Improved support for quoted identifiers.
 * Fixed creating patches with `snow app version create` when there are 2 or more existing patches on a version
+* Fixed bugs in source artifact mapping logic for native applications
 
 # v2.3.0
 
@@ -40,7 +41,6 @@
 * Fixed handling of stage path separators on Windows
 * Change to `external_access_integrations` in `snowflake.yml` now also triggers function replace
 * The `--info` callback returns info about configured feature flags.
-
 
 # v2.2.0